requirements.txt
streamlit==1.45.0
pandas==2.2.2
Pillow==11.0.0
PyMuPDF==1.25.1
openai==1.55.3
openpyxl==3.1.5
python-dotenv==1.0.1
